Removed uow_seq_number. Now uow_id is big serial.

+++ crabzilla-core/src/main/java/io/github/crabzilla/UnitOfWorkJournal.kt
@@ -2,8 +2,9 @@ package io.github.crabzilla
 
 import io.vertx.core.AsyncResult
 import io.vertx.core.Handler
+import java.math.BigInteger
 
 interface UnitOfWorkJournal {
 
-  fun append(unitOfWork: UnitOfWork, aHandler: Handler<AsyncResult<Int>>)
+  fun append(unitOfWork: UnitOfWork, aHandler: Handler<AsyncResult<BigInteger>>)
 }
